backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
batch-face==1.4.0
bitarray==2.9.2
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
cytoolz==0.12.3
decorator==5.1.1
decord==0.6.0
dill==0.3.8
eth-abi==5.1.0
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
facenet-pytorch==2.6.0
filelock==3.16.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
grad-cam==1.5.2
grpcio==1.66.1
h5py==3.11.0
hexbytes==1.2.1
huggingface-hub==0.24.7
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lmdb==1.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
more-itertools==10.2.0
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
parsimonious==0.10.0
pillow==10.2.0
pip==23.2.1
platformdirs==4.3.2
pooch==1.8.2
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pyunormalize==15.1.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rlp==4.0.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shap==0.44.1
slicer==0.0.7
setuptools==74.1.2
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.5
transformers==4.44.2
ttach==0.0.3
types-requests==2.32.0.20240907
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.93
ultralytics-thop==2.0.6
urllib3==2.2.3
web3==7.2.0
websockets==13.0.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
yarl==1.11.1
timm>=0.9.10
open_clip_torch
c2pa-python
cryptography
lightning
torchmetrics
lime